test(e2e): extend e2e metadata tests to google provider
